TODO:
nature repo, service.

connections...

HCs statuses models...

HostedServices for models...

users-roles-admins...

unit-tests...

comments...

+++ CompModels.CRUD.Services/Interfaces/IDataStorageFacadeService.cs
@@ -8,8 +8,8 @@ namespace CompModels.CRUD.Services.Interfaces
 {
     public interface IDataStorageFacadeService
     {
-        Task<int> AddCalculationRequestAsync(int compModelId, string inputJSON);
-        Task<int> GetCalculationRequestStatusIdAsync(int compModelId, int calculationRequestId, int userRequesterId);
+        Task<long> AddCalculationRequestAsync(int compModelId, string inputJSON);
+        Task<short> GetCalculationRequestStatusIdAsync(int compModelId, int calculationRequestId, int userRequesterId);
         Task<string> GetFinishedCalculationOutputAsync(int compModelId, int calculationRequestId, int userRequesterId);
     }
 }
